packages/aimms-web-backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==1.10.8

starlette>=0.36.3,<0.37.0
fastapi-crudrouter==0.8.6

# Database
SQLAlchemy==2.0.28
sqlalchemy-utils==0.41.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin bcrypt version for compatibility
python-dotenv==1.0.1

# HTTP client
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
certifi==2024.7.4
click==8.1.7
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
urllib3==2.2.2
yarl==1.9.4

# Logging
structlog==24.1.0

# Type hints
typing-extensions==4.9.0
annotated-types==0.6.0

PyJWT==2.4.0

websockets==10.3
wsproto==1.2.0

# Redis and Celery for background job processing
redis==5.0.1
celery==5.3.4

# Excel export
xlsxwriter>=3.0.0

# Email
aiosmtplib==3.0.2
Jinja2>=3.0.0
email-validator==2.1.0
